Krister Villarreal

+++ Episodio.cpp
@@ -14,7 +14,7 @@ Episodio::Episodio(){
     calificacion = 100;
 }
 
-Episodio::Episodio(string _titulo, int _temporada, double _calificacion){
+Episodio::Episodio(string _titulo, int _temporada, int _calificacion){
     titulo = _titulo;
     temporada = _temporada;
     calificacion = _calificacion;
@@ -30,7 +30,7 @@ void Episodio::setTemporada(int _temporada){
     temporada = _temporada;
 }
 
-void Episodio::setCalificacion(double _calificacion){ 
+void Episodio::setCalificacion(int _calificacion){ 
     calificacion = _calificacion;
 }
 
@@ -44,7 +44,7 @@ int Episodio::getTemporada(){
     return temporada;
 }
 
-double Episodio::getCalificacion(){
+int Episodio::getCalificacion(){
     return calificacion;
 }
